Work Experience component created
