+++ resumebuilder/src/App.js
@@ -2,6 +2,7 @@ import React from 'react';
 import './App.css';
 import Resume from './components/Resume'
 import applicant from './resume/applicant'
+import WorkExperience from './components/WorkExperience';
 
 function App() {
   return (
